SWREQ-BL017 - The BL can send a corresponding alert message of the electronic module to log and display. Also resolved comments on BLL-45 Peer Review BizElectronicModule, BizElectronicModuleState, BizElectronicModuleNormal, BizElectronicModuleAbnormal
git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk@191 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97

+++ PN022BLL/dal/dal/DalSimulationFile.h
@@ -84,7 +84,7 @@ namespace AtCor{
 									* @param[out] value4 Unsigned integer value from the fourth column in the file.
 									*
 									*/									
-									bool GetNextValues(signed int *value1, signed int *value2, signed int *value3, signed int *value4);
+									bool GetNextValues(unsigned int *value1, unsigned int *value2, unsigned int *value3, unsigned long *value4);
 
 									/**
 									* Resets the simulation source file to point to begining of file.
